refactor: remove StorageError
