+++ rust/src/lib.rs
@@ -105,7 +105,6 @@ pub use self::data_catalog::{get_data_catalog, DataCatalog, DataCatalogError};
 pub use self::delta::*;
 pub use self::partitions::*;
 pub use self::schema::*;
-pub use self::storage::StorageError;
 pub use object_store::{path::Path, Error as ObjectStoreError, ObjectMeta, ObjectStore};
 
 // convenience exports for consumers to avoid aligning crate versions
